Made even simple to add everything to one filegit add src/main/java/simplewallet/application/SimpleWalletRequestProcessor.java src/main/java/simplewallet/processor/SimpleWalletProcessor.java
